zenith-scrapper/requirements.txt
playwright
pydantic
pytest
loguru
pymysql
python-dotenv
prisma
